moves Lines in different location
