+++ src/main/java/ticTacToe/player/Player.java
@@ -2,7 +2,7 @@ package ticTacToe.player;
 
 import ticTacToe.game.Mark;
 import ticTacToe.grid.Grid;
-import ticTacToe.game.Lines;
+import ticTacToe.grid.Lines;
 import ticTacToe.ui.Ui;
 
 public interface Player {
